requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyPDF2
nltk
sqlalchemy
google-generativeai
